Prevent timer from going negative
Sometimes, `tickTimer()` (running on interval) is called an extra time
after the timer hits zero, but before `stopTimer()` (on timeout) has a
chance to do its thing. This makes the timer slip into the negatives and
displays `a second ago` as placeholder text. Modifying the timer to
avoid using a timeout function fixes the issue.