Fix issue where settings don't apply to message history (#261)
* No longer load settings after connection

The user isn't connected to chat until after chat history is loaded so
that the welcome message appears after the history. If settings are
loaded only after connecting, however, they won't apply to historical
messages. This change allows a setup where settings can be loaded prior
to history.

* Fix race condition

Not returning the promise from `loadSettings()` results in a race
condition where sometimes history still loads before settings.

* Apply server-saved settings even if not authed

With the new setup, the user is never authenticated with chat when
settings are fetched. This isn't a problem because authentication for
settings occurs independently.