Fix issue with events disappearing/reappearing
`contains()` doesn't consider an event removed if it's still in the DOM.
It isn't removed from the DOM until its remove animation is complete.
When coupled with `add()` checking if an event already exists before
adding it to the DOM, this can result in `replaceEvents()` simply
removing all events from the event bar.

Reworking `contains()` to use the `events` array to determine if an
event already exists instead fixes the issue.

This also disables animations when replacing events, which can look odd
when the new `PAIDEVENTS` payload contains old events.