Fix duplicate user messages on reconnect
An unresolved message has a hash that differs from its counterpart in
chat history as tracked by the server. This is because a message's
timestamp is used to generate its hash and client-side and server-side
timestamps differ. The system for preventing duplicate messages when
processing the chat history payload uses message hashes to determine if
a message already exists. As a result, if the user disconnects and
reconnects, their recent messages will be rendered to the DOM a second
time.

Modifying the message resolution system to update the timestamp of a
user's message after it's echoed by the server fixes the issue. After
resolution, a message's client-side timestamp will match its server-side
timestamp, as will its hash.