Fix 404 error when executing publish workflow
We recently converted the npm account that owned the dgg-chat-gui
package to an organization account. After this change, dgg-chat-gui now
belongs to the organization and is considered a scoped package, despite
not having a scope in its name. This resulted in an error when
attempting to publish the package.

Explicitly specifying `--access public` when publishing, as should be
done for public scoped packages, fixes the issue.